update environments folder to align with modularization/main
