+++ frontend/src/services/graphql/Environment/getTrustAccount.js
